Renamed classes and updated documentation
