+++ ojdbc-provider-gcp/src/main/java/oracle/jdbc/provider/gcp/configuration/GcpSecretManagerConfigurationProvider.java
@@ -35,52 +35,47 @@
  ** OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  ** SOFTWARE.
  */
-package oracle.jdbc.provider.gcp.objectstorage;
-
-import static oracle.jdbc.provider.parameter.Parameter.CommonAttribute.REQUIRED;
+package oracle.jdbc.provider.gcp.configuration;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
 
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
-
-import oracle.jdbc.provider.cache.CachedResourceFactory;
-import oracle.jdbc.provider.factory.Resource;
-import oracle.jdbc.provider.factory.ResourceFactory;
-import oracle.jdbc.provider.parameter.Parameter;
+import oracle.jdbc.driver.OracleConfigurationJsonProvider;
+import oracle.jdbc.provider.gcp.secrets.GcpSecretManagerFactory;
 import oracle.jdbc.provider.parameter.ParameterSet;
 
-public class GcpObjectStorageFactory implements ResourceFactory<InputStream> {
-  /** The secret version name for the secret */
-  public static final Parameter<String> PROJECT = Parameter.create(REQUIRED);
-  public static final Parameter<String> BUCKET = Parameter.create(REQUIRED);
-  public static final Parameter<String> OBJECT = Parameter.create(REQUIRED);
+/**
+ * A provider for JSON payload which contains configuration from GCP Secret
+ * Manager.
+ * See {@link #getJson(String)} for the spec of the JSON payload.
+ **/
+public class GcpSecretManagerConfigurationProvider extends OracleConfigurationJsonProvider {
 
-  private static final ResourceFactory<InputStream> INSTANCE = CachedResourceFactory
-      .create(new GcpObjectStorageFactory());
-
-  private GcpObjectStorageFactory() {
+  @Override
+  public String getType() {
+    return "gcpsecret";
   }
 
   /**
-   * Returns a singleton of {@code GcpVaultSecretFactory}.
+   * {@inheritDoc}
+   * <p>
+   * Returns the JSON payload stored in GCP Secret Manager secret.
+   * </p>
    * 
-   * @return a singleton of {@code GcpVaultSecretFactory}
+   * @param location resource name of the secret version (to obtain the resource
+   *                 name, click on "Actions" and "Copy resource name")
+   * @return JSON payload
    */
-  public static ResourceFactory<InputStream> getInstance() {
-    return INSTANCE;
+  @Override
+  public InputStream getJson(String location) throws SQLException {
+    Map<String, String> namedValues = new HashMap<>();
+    namedValues.put("secretVersionName", location);
+    ParameterSet parameterSet = GcpConfigurationParameters.getParser().parseNamedValues(namedValues);
+    return new ByteArrayInputStream(
+        GcpSecretManagerFactory.getInstance().request(parameterSet).getContent().getData().toByteArray());
   }
 
-  public Resource<InputStream> request(ParameterSet parameterSet)
-      throws IllegalStateException, IllegalArgumentException {
-    String projectName = parameterSet.getRequired(PROJECT);
-    String bucketName = parameterSet.getRequired(BUCKET);
-    String objectName = parameterSet.getRequired(OBJECT);
-
-    Storage storage = (Storage) StorageOptions.newBuilder().setProjectId(projectName).build().getService();
-    byte[] data = storage.readAllBytes(bucketName, objectName);
-    InputStream stream = new ByteArrayInputStream(data);
-    return Resource.createPermanentResource(stream, false);
-  }
 }
