+++ ojdbc-provider-gcp/src/main/resources/META-INF/services/oracle.jdbc.spi.OracleConfigurationJsonSecretProvider
@@ -1,1 +1,1 @@
-oracle.jdbc.provider.gcp.configuration.GcpJsonVaultSecretProvider
+oracle.jdbc.provider.gcp.configuration.GcpJsonSecretManagerProvider
