+++ ojdbc-provider-gcp/src/main/resources/META-INF/services/oracle.jdbc.spi.PasswordProvider
@@ -1,1 +1,1 @@
-oracle.jdbc.provider.gcp.resource.GcpVaultSecretPasswordProvider
+oracle.jdbc.provider.gcp.resource.GcpSecretManagerPasswordProvider
