+++ ojdbc-provider-gcp/src/main/resources/META-INF/services/oracle.jdbc.spi.UsernameProvider
@@ -1,1 +1,1 @@
-oracle.jdbc.provider.gcp.resource.GcpVaultSecretUsernameProvider
+oracle.jdbc.provider.gcp.resource.GcpSecretManagerUsernameProvider
